ocommerce/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-js-asset==1.2.2
django-mptt==0.12.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyMySQL==1.0.2
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
